+++ tst/babel/arrow.js
@@ -1,0 +1,11 @@
+/* eslint-disable */
+"use strict";
+
+/* This is an example of how babel exports a class */
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+let someArrow = () => 42;
+
+exports.default = someArrow;